refactor: extract migration logic into schema-migration.ts
Extract schema migration code from projects.ts into dedicated schema-migration.ts
module to keep core project logic clean and maintainable.

## Changes
- Create schema-migration.ts with:
  - isLegacySchema() — detects old groupId-keyed format
  - getRepoRemote() — auto-detects git remote URL
  - migrateLegacySchema() — converts old to new schema with channel merging

- Update projects.ts:
  - Remove migration functions (moved to schema-migration.ts)
  - Import isLegacySchema and migrateLegacySchema from schema-migration.ts
  - Keep core read/write logic clean

## Benefits
- Better separation of concerns
- Easier to understand projects.ts (core file I/O)
- Migration logic isolated and testable
- Reduces projects.ts from 520 lines to 344 lines
